squash update - lots of changes
write reg_direct

 squash update / rebase - ok. working. test03 to 4094 flip relay, and write register to control led

tidy, comment

ok. we are flipping between the test pattern and direct write

change reg direct value

control over mode

ok, direct control over mode, and direct register, and appears to work

tidy

better reporting

report more bits when trying to establish comms

fpga vector

comment

pack bitfield structure

remove mode, from relay test

ok, blink led using direct mode

tidy/ cleanup

update

+++ projects/dmm/include/4094.h
@@ -2,10 +2,12 @@
 
 #pragma once
 
+#include <stddef.h> // size_t
+
 
 void spi_4094_setup(uint32_t spi);
 uint8_t spi_4094_reg_write(uint32_t spi, uint8_t v);
 
 
-uint32_t spi_4094_reg_write_n(uint32_t spi, unsigned char *s, unsigned n);
+uint32_t spi_4094_reg_write_n(uint32_t spi, unsigned char *s, size_t n);
