+++ projects/dmm/src/4094.c
@@ -60,7 +60,7 @@ void spi_4094_setup(uint32_t spi)
   spi_init_master(
     spi,
     SPI_CR1_BAUDRATE_FPCLK_DIV_4,
-      // SPI_CR1_BAUDRATE_FPCLK_DIV_16,
+    // SPI_CR1_BAUDRATE_FPCLK_DIV_16,
     SPI_CR1_CPOL_CLK_TO_0_WHEN_IDLE,      // park to 0/lo == positive clok edge. park to 1 == negative clk edge.
     SPI_CR1_CPHA_CLK_TRANSITION_1,    // 1 == leading edge,  2 == falling edge
     SPI_CR1_DFF_8BIT,
@@ -110,7 +110,7 @@ uint8_t spi_4094_reg_write(uint32_t spi, uint8_t v)
 // think passing a unsigned char *s. is better.
 // can then call with &value.
 
-uint32_t spi_4094_reg_write_n(uint32_t spi, unsigned char *s, unsigned n)
+uint32_t spi_4094_reg_write_n(uint32_t spi, unsigned char *s, size_t n)
 {
   uint32_t ret = 0;
